services/web/requirements.txt
alembic==1.7.7
asn1crypto==1.5.1
attrs==21.4.0
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
freezegun==1.2.1
greenlet==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pg8000==1.29.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-pgsql==1.1.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
testing.common.database==2.0.3
testing.postgresql==1.3.0
tomli==2.0.1
Werkzeug==2.1.2
WTForms==3.0.1
